Put integration test in separate test file.
